Move functions that return JSX to separate files
