+++ packages/code-studio/src/settings/SettingsUtils.tsx
@@ -1,14 +1,11 @@
-import { DbNameValidator, TimeUtils } from '@deephaven/utils';
+import { DbNameValidator } from '@deephaven/utils';
 import {
   DateTimeColumnFormatter,
   IntegerColumnFormatter,
   DecimalColumnFormatter,
   TableColumnFormat,
-  TableUtils,
-  Formatter,
 } from '@deephaven/jsapi-utils';
 import Log from '@deephaven/log';
-import React, { ReactElement } from 'react';
 
 const log = Log.module('FormattingSectionContent');
 
@@ -108,82 +105,6 @@ export function isFormatRuleValidForSave(rule: FormatterItem): boolean {
   );
 }
 
-export function renderTimeZoneOptions(): JSX.Element[] {
-  const options = TimeUtils.TIME_ZONES.map(timeZone => {
-    const { label, value } = timeZone;
-    return (
-      <option value={value} key={value}>
-        {label}
-      </option>
-    );
-  });
-
-  return options;
-}
-
-export function renderColumnTypeOptions(): ReactElement {
-  const columnTypesArray = [
-    { value: TableUtils.dataType.DATETIME, label: 'DateTime' },
-    { value: TableUtils.dataType.DECIMAL, label: 'Decimal' },
-    { value: TableUtils.dataType.INT, label: 'Integer' },
-  ];
-
-  const columnTypeOptions = columnTypesArray.map(item => (
-    <option key={`key-columnType-${item.value}`} value={item.value}>
-      {item.label}
-    </option>
-  ));
-
-  return (
-    <>
-      <option key="key-columnType-placeholder" disabled value="">
-        Select Type
-      </option>
-      {columnTypeOptions}
-    </>
-  );
-}
-
-export function renderDateTimeOptions(
-  timestamp: Date,
-  timeZone: string,
-  showTimeZone: boolean,
-  showTSeparator: boolean,
-  isGlobalOptions: boolean,
-  legacyGlobalFormat?: string
-): ReactElement[] {
-  const formatter = new Formatter([], {
-    timeZone,
-    showTimeZone,
-    showTSeparator,
-  });
-  const formats = isGlobalOptions
-    ? DateTimeColumnFormatter.getGlobalFormats(showTimeZone, showTSeparator)
-    : DateTimeColumnFormatter.getFormats(showTimeZone, showTSeparator);
-
-  if (legacyGlobalFormat != null && !formats.includes(legacyGlobalFormat)) {
-    formats.unshift(legacyGlobalFormat);
-  }
-
-  return formats.map(formatString => {
-    const format = DateTimeColumnFormatter.makeFormat(
-      '',
-      formatString,
-      DateTimeColumnFormatter.TYPE_GLOBAL
-    );
-    return (
-      <option value={formatString} key={formatString}>
-        {formatter.getFormattedString(
-          timestamp,
-          TableUtils.dataType.DATETIME,
-          '',
-          format
-        )}
-      </option>
-    );
-  });
-}
-
 export default {
   focusFirstInputInContainer,
   isSameOptions,
@@ -193,7 +114,4 @@ export default {
   isValidFormat,
   removeFormatRuleExtraProps,
   isFormatRuleValidForSave,
-  renderTimeZoneOptions,
-  renderColumnTypeOptions,
-  renderDateTimeOptions,
 };
